Support float values for buienradar

+++ draw/buienradar.py
@@ -3,7 +3,7 @@ from math import sin, cos, pi, pow
 def draw_buienradar_chart(context, x, y, w, h, data):
     context.draw.rectangle((x, y+h, x+w, y))
     n = len(data) - 1
-    pts = [(x + int(1.0 * w * i / n), y+h - int(1.0 * h * int(p[1]) / 255)) for i, p in enumerate(data)]
+    pts = [(x + int(1.0 * w * i / n), y+h - int(1.0 * h * float(p[1]) / 255)) for i, p in enumerate(data)]
     context.draw.line(pts, width=3, fill=context.black)
     draw_raindrop(context, x + w / 2, y + h / 3, 5)
